% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{atencao}
\alias{atencao}
\title{Sugere atos que merecem atenção}
\usage{
atencao(conferir, nomes, gabarito)
}
\arguments{
\item{conferir}{um vetor com o texto a ser conferido}

\item{nomes}{o vetor com os nomes ou ids dos textos a serem conferidos}

\item{gabarito}{vetor com o qual as respostas devem ser próximas}
}
\value{
Um data.frame com um vetor dizendo se a observação merece atenção
        e outro vetor com o nome da observação sugerida
}
\description{
Sugere atos que merecem atenção
}
\examples{
#Sem exemplo
}
