% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auxiliar}
\alias{auxiliar}
\title{Lista de tags usadas nos scripts.}
\format{Um data frame com 1 campo e 4 observações:
\describe{
  \item{tag}{nome da tag}
}}
\source{
W3
}
\usage{
auxiliar
}
\description{
Um data frame com as tags usadas em loops dos scripts
}
\keyword{datasets}
