% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{conteudo_orgao}
\alias{conteudo_orgao}
\title{Conteudo de um orgao}
\usage{
conteudo_orgao(orgao_alvo, conteudo, orgaos, arquivo, consulta = FALSE)
}
\arguments{
\item{orgao_alvo}{nome do orgao procurado}

\item{conteudo}{vetor de texto com conteudo do DOU}

\item{orgaos}{vetor de texto com orgaos publicados naquele dia e secao}

\item{arquivo}{arquivo com informacoes de secao e data do DOU}

\item{consulta}{Se TRUE retorna o nome do orgao_alvo como esta no DOU}
}
\value{
texto com as publicacoes do orgao alvo
}
\description{
Conteudo de um orgao
}
