% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{converter_pdf}
\alias{converter_pdf}
\title{Converter PDFs em TXTs}
\usage{
converter_pdf(pdf_dir, txt_dir)
}
\arguments{
\item{pdf_dir}{Pasta onde estao os arquivos PDF}

\item{txt_dir}{Pasta onde ficarao os arquivos TXT}
}
\description{
Converter PDFs em TXTs
}
