% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{criar_indices}
\alias{criar_indices}
\title{Criar indices}
\usage{
criar_indices(vetor_inicios)
}
\arguments{
\item{vetor_inicios}{Vetor de indices de inicios}
}
\value{
uma lista com os indices entre cada elemento de \code{vetor_inicios}
}
\description{
Criar indices
}
