% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_tabela_ato.R
\name{criar_tabela_ato}
\alias{criar_tabela_ato}
\title{Tabela Ato}
\usage{
criar_tabela_ato(lista_de_atos)
}
\arguments{
\item{lista_de_atos}{Lista de atos retirados de um dia do DOU com a função `pega_normas_dou()`}
}
\value{
Uma tibble com os fa tabela Ato no banco de dados
}
\description{
Tabela Ato
}
\examples{
# Sem exemplo

}
