% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_tabela_item.R
\name{criar_tabela_item}
\alias{criar_tabela_item}
\title{Tabela Item}
\usage{
criar_tabela_item(lista_de_atos)
}
\arguments{
\item{lista_de_atos}{}
}
\value{
Uma tabela com os mesmos campos da tabela Item no Banco de Dados
}
\description{
Tabela Item
}
