% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_orgaos}
\alias{dic_orgaos}
\title{Dicionário dos orgaos}
\format{Um data frame com 2 campos e 34 observações:
\describe{
  \item{SGL_ORGAO}{Sigla do orgao}
  \item{DES_ORGAO}{Descricao do orgao}
}}
\source{
interno
}
\usage{
dic_orgaos
}
\description{
Um data frame com as siglas e descricoes dos orgao
}
\keyword{datasets}
