% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_tipos}
\alias{dic_tipos}
\title{Dicionário dos tipos de normas}
\format{Um data frame com 2 campos e 30 observações:
\describe{
  \item{SGL_TIPO}{Sigla do tipo}
  \item{DES_TIPO}{Descrição do tipo}
}}
\source{
interno
}
\usage{
dic_tipos
}
\description{
Um data frame com os codigos dos tipos de normas
}
\keyword{datasets}
