% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{download_dou}
\alias{download_dou}
\title{Download dos PDFs do Diario Oficial da Uniao}
\usage{
download_dou(data, dest_dir)
}
\arguments{
\item{data}{Data que sera baixada}

\item{pdf_dir}{Pasta onde estao os arquivos PDF}
}
\description{
Download dos PDFs do Diario Oficial da Uniao
}
