% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{eliminar_quebras}
\alias{eliminar_quebras}
\title{Eliminar quebras indesejadas}
\usage{
eliminar_quebras(texto)
}
\arguments{
\item{texto}{Texto que tera as quebras de linhas erradas eliminadas}
}
\value{
O mesmo texto sem as quebras indesejadas
}
\description{
Eliminar quebras indesejadas
}
\examples{
eliminar_quebras('Um texto com \\n2 quebras de linha\\n')

}
