% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{\%em\%}
\alias{\%em\%}
\title{Qual faixa de b (limite) contem a (conteudo)?}
\usage{
conteudo \%em\% limite
}
\arguments{
\item{conteudo}{vetor a ser analisado}

\item{limite}{vetor com limites}
}
\value{
Em qual intervalo de \code{limite} encontra-se cada elemento de \code{conteudo}.
}
\description{
Qual faixa de b (limite) contem a (conteudo)?
}
