% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{estruturar_normas}
\alias{estruturar_normas}
\title{Tabela para Validação na Aplicação}
\usage{
estruturar_normas(lista_de_normas)
}
\arguments{
\item{lista_de_normas}{Lista de normas retornada pela função \code{pegar_normas_dou}.}
}
\value{
tabela com informações que precisam ser validadas na aplicação
}
\description{
Tabela para Validação na Aplicação
}
