% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_id.R
\name{gerar_id}
\alias{gerar_id}
\title{Adiciona IDs ao data.frame}
\usage{
gerar_id(df, anterior)
}
\arguments{
\item{lista_de_atos}{}
}
\value{
O mesmo df precedido por novos IDs
}
\description{
Adiciona IDs ao data.frame
}
