% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operador_em.R
\name{\%em\%}
\alias{\%em\%}
\title{Qual faixa de b contem a?}
\usage{
conteudo \%em\% limite
}
\arguments{
\item{a}{um vetor}

\item{b}{um vetor}
}
\value{
Em qual \code{b} encontra-se cada elemento de \code{a}.
}
\description{
Qual faixa de b contem a?
}
\examples{
#Sem exemplo
}
