% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{inicios_atos}
\alias{inicios_atos}
\title{Encontrar inicio dos atos no texto}
\usage{
inicios_atos(texto)
}
\arguments{
\item{texto}{Texto em que as normas serao buscadas}
}
\value{
vetor  numerico com inicio dos atos
}
\description{
Encontrar inicio dos atos no texto
}
