% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{levsim}
\alias{levsim}
\title{Aplicacao de algoritmo de levenshtein entre vetor e valor escalar}
\usage{
levsim(str1, str2)
}
\arguments{
\item{str1}{primeira string}

\item{str2}{segunda string}
}
\value{
vetor de distancias
}
\description{
Aplicacao de algoritmo de levenshtein entre vetor e valor escalar
}
