% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa_atributos.R
\name{limpa_atributos}
\alias{limpa_atributos}
\title{Limpa atributos de um html}
\usage{
limpa_atributos(html)
}
\arguments{
\item{html}{um vetor com html}
}
\value{
O mesmo \code{html} sem os atributos em todas as tags
}
\description{
Limpa atributos de um html
}
\examples{
#Sem exemplo
}
