% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{limpar_atributos}
\alias{limpar_atributos}
\title{Faz download de todas as páginas do DOU da seção e dia escolhidos}
\usage{
limpar_atributos(html)
}
\arguments{
\item{html}{um vetor com html}

\item{data}{Data do diário que será baixada. Deve ser fornecida no formato "%d/%m/%Y"}

\item{diario}{Número da seção do diário (1, 2 ou 3)}
}
\value{
nada. Seu objetivo são efeitos colaterais (donwload)

O mesmo \code{html} sem os atributos em todas as tags
}
\description{
Faz download de todas as páginas do DOU da seção e dia escolhidos
}
\examples{
#Sem exemplo

#Sem exemplo
}
