% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{limpar_texto}
\alias{limpar_texto}
\title{Remover erros de interpretacao (parse) do texto}
\usage{
limpar_texto(texto)
}
\arguments{
\item{texto}{O texto que sera limpo}
}
\value{
O mesmo texto limpo
}
\description{
Remover erros de interpretacao (parse) do texto
}
