% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipla_para_individualizada.R
\name{multipla_para_individualizada}
\alias{multipla_para_individualizada}
\title{Cria um objeto para cada portaria em objeto de portarias multiplas}
\usage{
multipla_para_individualizada(portaria)
}
\arguments{
\item{portaria}{um vetor o texto da portaria multipla}
}
\value{
um vetor com uma portaria em cada elemento
}
\description{
Cria um objeto para cada portaria em objeto de portarias multiplas
}
\examples{
#Sem exemplo
}
