% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsear_e_escrever.R
\name{parsear_e_escrever}
\alias{parsear_e_escrever}
\title{Parsear normas em uma sessao do DOU e escrever na base}
\usage{
parsear_e_escrever(conexao, pastas, quais = seq_len(10), debug = FALSE)
}
\arguments{
\item{conexao}{Uma conexao}

\item{pastas}{Vetor com pastas em que se encontram as paginas txt de uma unidade do DOU}

\item{quais}{ID do ultimo ato escrito na base. Se nao existe ID = 0}

\item{debug}{Debugando?}
}
\description{
Parsear normas em uma sessao do DOU e escrever na base
}
