% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_cargo.R
\name{pega_cargo}
\alias{pega_cargo}
\title{Pega cargo da autoridade que assintou o ato}
\usage{
pega_cargo(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
O cargo da autoridade que assintou um \code{ato}.
}
\description{
Pega cargo da autoridade que assintou o ato
}
\examples{
#Sem exemplo
}
