% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pega_dados_BP}
\alias{pega_dados_BP}
\title{Pega todos os dados de todos os atos de uma edicao do Boletim de Pessoal (pdf)}
\usage{
pega_dados_BP(arquivo, debug = FALSE)
}
\arguments{
\item{arquivo}{O caminho do arquivo (.pdf) de uma edicao do Boletim de Pessoal}
}
\value{
Uma tabela com todos os dados extraidos do DOU.
}
\description{
Pega todos os dados de todos os atos de uma edicao do Boletim de Pessoal (pdf)
}
\examples{
# Sem exemplo


}
