% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_data.R
\name{pega_data}
\alias{pega_data}
\title{Pega data do nome de um ato}
\usage{
pega_data(vetor)
}
\arguments{
\item{vetor}{um vetor com o nome de um ato}
}
\value{
A data de \code{vetor} no formato dia DE MES_POR_EXTENSO DE ANO
}
\description{
Pega data do nome de um ato
}
\examples{
#Sem exemplo
}
