% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_limites_orgaos.R
\name{pega_limites_orgaos}
\alias{pega_limites_orgaos}
\title{Pega o limite dos orgaos de um ministerio}
\usage{
pega_limites_orgaos(pagina)
}
\arguments{
\item{pagina}{um vetor com todo o conteudo de um dia do DOU}
}
\value{
Em qual linha (elemento) de \code{pagina} estao os orgaos de um ministerio
}
\description{
Pega o limite dos orgaos de um ministerio
}
\examples{
#Sem exemplo
}
