% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_normas_dou.R
\name{pega_normas_dou}
\alias{pega_normas_dou}
\title{Pega todos os dados de todos os atos de um dia do DOU (txt)}
\usage{
pega_normas_dou(arquivos, debug = FALSE)
}
\arguments{
\item{arquivos}{um vetor com os caminhos dos arquivos (.txt) de um dia do DOU}
}
\value{
Uma lista com todas as normas extraidas do DOU.
}
\description{
Pega todos os dados de todos os atos de um dia do DOU (txt)
}
\examples{
#Sem exemplo
}
