% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_numero.R
\name{pega_numero}
\alias{pega_numero}
\title{Pega numero dos atos}
\usage{
pega_numero(vetor)
}
\arguments{
\item{vetor}{um vetor com o nome de um ato}
}
\value{
A data de \code{vetor} no formato dia DE MES_POR_EXTENSO DE ANO
}
\description{
Pega numero dos atos
}
\examples{
#Sem exemplo
}
