% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_resumo.R
\name{pega_resumo}
\alias{pega_resumo}
\title{Pega numero dos vetors}
\usage{
pega_resumo(vetor)
}
\arguments{
\item{vetor}{um vetor com o nome de um vetor}
}
\value{
A data de \code{vetor} no formvetor dia DE MES_POR_EXTENSO DE ANO
}
\description{
Pega numero dos vetors
}
\examples{
#Sem exemplo
}
