% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_tipo.R
\name{pega_tipo}
\alias{pega_tipo}
\title{Pega tipo dos ato}
\usage{
pega_tipo(vetor, retorno = "txt")
}
\arguments{
\item{vetor}{um vetor com o conteudo de um ato}
}
\value{
O tipo do ato de \code{vetor}.
}
\description{
Pega tipo dos ato
}
\examples{
#Sem exemplo
}
