% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_cargo}
\alias{pegar_cargo}
\title{Pega cargo da autoridade que assintou o ato}
\usage{
pegar_cargo(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
O cargo da autoridade que assintou o \code{ato}.
}
\description{
Pega cargo da autoridade que assintou o ato
}
\examples{
#Sem exemplo
}
