% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_limites_orgaos}
\alias{pegar_limites_orgaos}
\title{Pegar o limite dos orgaos de um ministerio}
\usage{
pegar_limites_orgaos(pagina)
}
\arguments{
\item{pagina}{um vetor com todo o conteudo de um dia do DOU}
}
\value{
Em qual linha (elemento) de \code{pagina} estao os orgaos de um ministerio
}
\description{
Pegar o limite dos orgaos de um ministerio
}
