% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_normas_dou}
\alias{pegar_normas_dou}
\title{Pega todos os dados de todos os atos de um dia do DOU (txt)}
\usage{
pegar_normas_dou(arquivos, orgao_alvo, encodificacao = "latin1")
}
\arguments{
\item{arquivos}{um vetor com os caminhos dos arquivos (.txt) de um dia do DOU}

\item{orgao_alvo}{orgão cujas normas seram buscadas}

\item{encodificacao}{encodificação dos arquivos}
}
\value{
Uma lista com todas as normas extraidas do DOU e algumas mata-informações
}
\description{
Pega todos os dados de todos os atos de um dia do DOU (txt)
}
