% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_numero}
\alias{pegar_numero}
\title{Pega numero dos atos}
\usage{
pegar_numero(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
A data de \code{vetor} no formato dia DE MES_POR_EXTENSO DE ANO
}
\description{
Pega numero dos atos
}
\examples{
# Sem exemplo
}
