% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_resumo}
\alias{pegar_resumo}
\title{Pegar resumo do ato}
\usage{
pegar_resumo(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
A data de \code{vetor} no formvetor dia DE MES_POR_EXTENSO DE ANO
}
\description{
Pegar resumo do ato
}
