% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_sigla_orgao}
\alias{pegar_sigla_orgao}
\title{Pegar Sigla do Orgao}
\usage{
pegar_sigla_orgao(nome_orgao)
}
\arguments{
\item{nome_orgao}{Nome do orgao cuja sigla sera buscada}
}
\value{
A sigla do orgao passado como argumento
}
\description{
Pegar Sigla do Orgao
}
