% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_titulo}
\alias{pegar_titulo}
\title{Título do Ato}
\usage{
pegar_titulo(ato)
}
\arguments{
\item{ato}{Vetor de texto com conteudo da norma legal}
}
\value{
Título de um ato de seu corpo de texto
}
\description{
Título do Ato
}
