% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procura_html.R
\name{procura_html}
\alias{procura_html}
\title{Procura um termo em uma tag de um html}
\usage{
procura_html(html, tag, termo, letra = FALSE, value = FALSE)
}
\arguments{
\item{tag}{um vetor com o nome de um ato}

\item{termo}{um vetor com o nome de um ato}

\item{hmtl}{um vetor com o nome de um ato}
}
\value{
O enésimo item do texto com correspondência do \code{termo} na \code{tag} do \code{html}
}
\description{
Procura um termo em uma tag de um html
}
\examples{
#Sem exemplo
}
