% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procura_inicio.R
\name{procura_inicio}
\alias{procura_inicio}
\title{Procura um termo no inicio de uma string}
\usage{
procura_inicio(vetor, termo)
}
\arguments{
\item{vetor}{um vetor de um dia do DOU}

\item{termo}{termo a ser buscado}
}
\value{
O indice do \code{vetor} em que o termo se encontra no inicio da string
}
\description{
Procura um termo no inicio de uma string
}
\examples{
#Sem exemplo
}
