% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{procurar_inicio}
\alias{procurar_inicio}
\title{Procura um termo no inicio de uma string}
\usage{
procurar_inicio(vetor, termo)
}
\arguments{
\item{vetor}{um vetor de texto}

\item{termo}{termo a ser buscado}
}
\value{
O índice do \code{vetor} em que o \code{termo} se encontra no inicio da string
}
\description{
Procura um termo no inicio de uma string
}
