% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{remover_tags}
\alias{remover_tags}
\title{Elimina as tags de um documento html}
\usage{
remover_tags(html)
}
\arguments{
\item{html}{um vetor com html}
}
\value{
O mesmo \code{html} sem as tags
}
\description{
Elimina as tags de um documento html
}
\examples{
#Sem exemplo
}
