% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_para_html.R
\name{texto_para_html}
\alias{texto_para_html}
\title{Transformar Texto em Parágrafos de HTML}
\usage{
texto_para_html(texto)
}
\arguments{
\item{texto}{}
}
\description{
Transformar Texto em Parágrafos de HTML
}
