% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tira_estilo.R
\name{tira_estilo}
\alias{tira_estilo}
\title{Elimina as tags de estilo um documento html}
\usage{
tira_estilo(html)
}
\arguments{
\item{html}{um vetor com html}
}
\value{
O mesmo \code{html} sem as tags de estilo
}
\description{
Elimina as tags de estilo um documento html
}
\examples{
#Sem exemplo
}
