% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{converter_pdf}
\alias{converter_pdf}
\title{Converter PDFs em TXTs}
\usage{
converter_pdf(data, secao, dir_pdf = "pdf", dest_dir = "txt")
}
\arguments{
\item{data}{Data do DOU}

\item{secao}{Secao do DOU}

\item{pdf_dir}{Pasta onde ficarao os arquivos PDF}

\item{txt_dir}{Pasta onde ficarao os arquivos TXT}
}
\description{
Converter PDFs em TXTs
}
