% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{criar_tabela_app}
\alias{criar_tabela_app}
\title{Tabela para Validação na Aplicação}
\usage{
criar_tabela_app(lista_de_atos)
}
\arguments{
\item{lista_de_atos}{}
}
\value{
tabela com informações que precisam ser validadas na aplicação
}
\description{
Tabela para Validação na Aplicação
}
\examples{
# Sem exemplos
}
