% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{gerar_id}
\alias{gerar_id}
\title{Adiciona IDs ao data.frame}
\usage{
gerar_id(df, anterior)
}
\arguments{
\item{df}{Um data.frame}

\item{anterior}{Caminho do RDS com a tabela anterior de cujos ID devem se seguir}
}
\value{
O mesmo df precedido por novos IDs
}
\description{
Adiciona IDs ao data.frame
}
\examples{
# Sem exemplo

}
