% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{multipla_para_individualizada}
\alias{multipla_para_individualizada}
\title{Cria um objeto para cada norma em objeto de normas multiplas}
\usage{
multipla_para_individualizada(norma)
}
\arguments{
\item{norma}{um vetor o texto da norma multipla}
}
\value{
um vetor com uma norma em cada elemento
}
\description{
Cria um objeto para cada norma em objeto de normas multiplas
}
