% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{parsear_e_escrever}
\alias{parsear_e_escrever}
\title{Parsear normas em uma sessao do DOU e escrever na base}
\usage{
parsear_e_escrever(conexao, pastas, debug = FALSE)
}
\arguments{
\item{conexao}{Uma conexao}

\item{pastas}{Vetor com pastas em que se encontram as paginas txt de uma unidade do DOU}

\item{debug}{Debugando?}
}
\description{
Parsear normas em uma sessao do DOU e escrever na base
}
