% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_data}
\alias{pegar_data}
\title{Pegar data do nome de um ato}
\usage{
pegar_data(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
A data do \code{ato}
}
\description{
Pegar data do nome de um ato
}
