% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_normas_dou}
\alias{pegar_normas_dou}
\title{Pega todos os dados de todos os atos de um dia do DOU (txt)}
\usage{
pegar_normas_dou(arquivos, debug = FALSE)
}
\arguments{
\item{arquivos}{um vetor com os caminhos dos arquivos (.txt) de um dia do DOU}

\item{debug}{A função está sendo debugada?}
}
\value{
Uma lista com todas as normas extraidas do DOU e algumas mata-informações
}
\description{
Pega todos os dados de todos os atos de um dia do DOU (txt)
}
