% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_numero}
\alias{pegar_numero}
\title{Pegar numero dos atos}
\usage{
pegar_numero(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
O numero de \code{ato} com 8 digitos (completados com 0)
}
\description{
Pegar numero dos atos
}
