% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_pagina}
\alias{pegar_pagina}
\title{Número da Página do ato}
\usage{
pegar_pagina(ato, arquivos)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}

\item{arquivos}{vetor com caminho dos arquivos em que o ato pode estar}
}
\value{
Página de cada ato
}
\description{
Número da Página do ato
}
\examples{
# Sem exemplo

}
