% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_pagina.R
\name{pegar_pagina}
\alias{pegar_pagina}
\title{Número da Página do ato}
\usage{
pegar_pagina(ato, arquivos)
}
\arguments{
\item{ato}{}

\item{arquivos}{}
}
\value{
Página de cada ato
}
\description{
Número da Página do ato
}
