% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_infos.R
\name{pegar_tipo}
\alias{pegar_tipo}
\title{Pegar tipo dos ato}
\usage{
pegar_tipo(ato)
}
\arguments{
\item{ato}{um vetor com o conteudo de um ato}
}
\value{
O tipo do ato de \code{vetor}.
}
\description{
Pegar tipo dos ato
}
