% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{texto_para_html}
\alias{texto_para_html}
\title{Transformar Texto em Parágrafos de HTML}
\usage{
texto_para_html(texto)
}
\arguments{
\item{texto}{Um vetor com texto}
}
\value{
O mesmo texto com as quebras de linha substituídas por tags de parágrafos \code{<p>}.
        Adiciona um cabeçalho.
}
\description{
Transformar Texto em Parágrafos de HTML
}
\examples{
# Sem exemplo

}
