% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{texto_para_html}
\alias{texto_para_html}
\title{Transformar Texto em Parágrafos de HTML}
\usage{
texto_para_html(texto, orgao, cabecalho)
}
\arguments{
\item{texto}{Um vetor com texto}

\item{orgao}{Nome do orgao que sera inserido no texto}

\item{cabecalho}{Cabecalho que sera incluido no inicio do texto}
}
\value{
O mesmo texto com as quebras de linha substituídas por tags de parágrafos \code{<p>}.
        Adiciona um cabeçalho.
}
\description{
Transformar Texto em Parágrafos de HTML
}
