% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{tirar_estilo}
\alias{tirar_estilo}
\title{Elimina as tags de estilo um documento html}
\usage{
tirar_estilo(html)
}
\arguments{
\item{html}{um vetor com html}
}
\value{
O mesmo \code{html} sem as tags de estilo \code{<style>},
        da Office \code{<o:*>} ou \code{<v:*>}
}
\description{
Elimina as tags de estilo um documento html
}
\examples{
#Sem exemplo
}
